+++ components/zcash_address/src/test_vectors.rs
@@ -8,6 +8,7 @@ use {
         unified::{
             self,
             address::{test_vectors::TEST_VECTORS, Receiver},
+            Item,
         },
         Network, ToAddress, ZcashAddress,
     },
@@ -36,6 +37,7 @@ fn unified() {
                     data: data.to_vec(),
                 })
             }))
+            .map(Item::Data)
             .collect();
 
         let expected_addr = ZcashAddress::from_unified(Network::Main, unified::Address(receivers));