+++ components/zcash_address/CHANGELOG.md
@@ -7,6 +7,25 @@ and this library adheres to Rust's notion of
 
 ## [Unreleased]
 
+### Added
+- `zcash_address::unified`:
+  - `Address::receivers`
+  - `DataTypecode`
+  - `MetadataTypecode`
+  - `Item`
+  - `MetadataItem`
+  - `Address::receivers`
+
+### Changed
+- `zcash_address::unified`:
+  - `Typecode` has changed. Instead of having a variant for each receiver type,
+    it now has two variants, `Typecode::Data` and `Typecode::Metadata`.
+
+### Removed
+- `zcash_address::unified::Container::items` Preference order is only
+  significant when considering unified address receivers; use
+  `Address::receivers` instead.
+
 ## [0.3.2] - 2024-03-06
 ### Added
 - `zcash_address::convert`: